+++ fuc/cli/ngs_fq2bam.py
@@ -33,7 +33,6 @@ epilog = f"""
   ref.fa \\
   output_dir \\
   "-q queue_name -pe pe_name 10" \\
-  "-q queue_name" \\
   "-Xmx15g -Xms15g" \\
   1.vcf 2.vcf 3.vcf \\
   --thread 10
@@ -44,7 +43,6 @@ epilog = f"""
   ref.fa \\
   output_dir \\
   "-l h='node_A|node_B' -pe pe_name 10" \\
-  "-l h='node_A|node_B'" \\
   "-Xmx15g -Xms15g" \\
   1.vcf 2.vcf 3.vcf \\
   --thread 10
@@ -72,18 +70,9 @@ def create_parser(subparsers):
         help='Output directory.'
     )
     parser.add_argument(
-        'qsub1',
+        'qsub',
         type=str,
-        help="SGE resoruce to request with qsub for read alignment \n"
-             "and sorting. Since both tasks support multithreading, \n"
-             "it is recommended to speicfy a parallel environment (PE) \n"
-             "to speed up the process (also see --thread)."
-    )
-    parser.add_argument(
-        'qsub2',
-        type=str,
-        help="SGE resoruce to request with qsub for the rest of the \n"
-             "tasks, which do not support multithreading."
+        help='SGE resoruce to request for qsub.'
     )
     parser.add_argument(
         'java',
@@ -117,6 +106,12 @@ def create_parser(subparsers):
         help="Sequencing platform (default: 'Illumina')."
     )
     parser.add_argument(
+        '--job',
+        metavar='TEXT',
+        type=str,
+        help='Job submission ID for SGE.'
+    )
+    parser.add_argument(
         '--force',
         action='store_true',
         help='Overwrite the output directory if it already exists.'
@@ -146,12 +141,58 @@ def main(args):
     else:
         remove = 'rm'
 
+    java_shared = f'-XX:ParallelGCThreads={args.thread} -XX:ConcGCThreads={args.thread}'
+
     for i, r in df.iterrows():
         with open(f'{args.output}/shell/S1-{r.Name}.sh', 'w') as f:
 
             ###########
             # BWA-MEM #
             ###########
+
+            command1 = 'bwa mem'
+            command1 += f' -M -R $group -t {args.thread}'
+            command1 += f' {args.fasta} {r.Read1} {r.Read2} |'
+            command1 += f' samtools sort -@ {args.thread}'
+            command1 += f' -o {args.output}/temp/{r.Name}.sorted.bam -'
+
+            ##################
+            # MarkDuplicates #
+            ##################
+
+            command2 = 'gatk MarkDuplicates'
+            command2 += f' --QUIET'
+            command2 += f' --java-options "{java_shared} {args.java}"'
+            command2 += f' -I {args.output}/temp/{r.Name}.sorted.bam'
+            command2 += f' -O {args.output}/temp/{r.Name}.sorted.markdup.bam'
+            command2 += f' -M {args.output}/temp/{r.Name}.metrics'
+
+            ####################
+            # BaseRecalibrator #
+            ####################
+
+            command3 = 'gatk BaseRecalibrator'
+            command3 += f' --QUIET'
+            command3 += f' --java-options "{java_shared} {args.java}"'
+            command3 += f' -R {args.fasta}'
+            command3 += f' -I {args.output}/temp/{r.Name}.sorted.markdup.bam'
+            command3 += f' -O {args.output}/temp/{r.Name}.table'
+            command3 += ' ' + ' '.join([f'--known-sites {x}' for x in args.vcf])
+            if args.bed is not None:
+                command3 += f' -L {args.bed}'
+
+            #############
+            # ApplyBQSR #
+            #############
+
+            command4 = 'gatk ApplyBQSR'
+            command4 += f' --QUIET'
+            command4 += f' --java-options "{java_shared} {args.java}"'
+            command4 += f' -bqsr {args.output}/temp/{r.Name}.table'
+            command4 += f' -I {args.output}/temp/{r.Name}.sorted.markdup.bam'
+            command4 += f' -O {args.output}/{r.Name}.sorted.markdup.recal.bam'
+            if args.bed is not None:
+                command4 += f' -L {args.bed}'
 
             f.write(
 f"""#!/bin/bash
@@ -166,68 +207,26 @@ barcode=`echo "$first" | awk -F " " '{{print $2}}' | awk -F ":" '{{print $4}}'`
 group="@RG\\tID:$flowcell\\tPU:$flowcell.$barcode\\tSM:{r.Name}\\tPL:{args.platform}\\tLB:{r.Name}"
 
 # Align and sort seuqnece reads. Assign read group as well.
-bwa mem -M -R $group -t {args.thread} {args.fasta} {r.Read1} {r.Read2} | samtools sort -@ {args.thread} -o {args.output}/temp/{r.Name}.sorted.bam -
-""")
+{command1}
 
-        with open(f'{args.output}/shell/S2-{r.Name}.sh', 'w') as f:
-
-            ##################
-            # MarkDuplicates #
-            ##################
-
-            command1 = 'gatk MarkDuplicates'
-            command1 += f' --QUIET'
-            command1 += f' --java-options "{args.java}"'
-            command1 += f' -I {args.output}/temp/{r.Name}.sorted.bam'
-            command1 += f' -O {args.output}/temp/{r.Name}.sorted.markdup.bam'
-            command1 += f' -M {args.output}/temp/{r.Name}.metrics'
-
-            ####################
-            # BaseRecalibrator #
-            ####################
-
-            command2 = 'gatk BaseRecalibrator'
-            command2 += f' --QUIET'
-            command2 += f' --java-options "{args.java}"'
-            command2 += f' -R {args.fasta}'
-            command2 += f' -I {args.output}/temp/{r.Name}.sorted.markdup.bam'
-            command2 += f' -O {args.output}/temp/{r.Name}.table'
-            command2 += ' ' + ' '.join([f'--known-sites {x}' for x in args.vcf])
-
-            if args.bed is not None:
-                command2 += f' -L {args.bed}'
-
-            #############
-            # ApplyBQSR #
-            #############
-
-            command3 = 'gatk ApplyBQSR'
-            command3 += f' --QUIET'
-            command3 += f' --java-options "{args.java}"'
-            command3 += f' -bqsr {args.output}/temp/{r.Name}.table'
-            command3 += f' -I {args.output}/temp/{r.Name}.sorted.markdup.bam'
-            command3 += f' -O {args.output}/{r.Name}.sorted.markdup.recal.bam'
-
-            if args.bed is not None:
-                command3 += f' -L {args.bed}'
-
-            f.write(
-f"""#!/bin/bash
-
-# Activate conda environment.
-source activate {api.common.conda_env()}
+if [ $? -eq 0 ]; then
+    echo "Successfully finished aligning and sorting seuqnece reads"
+else
+    echo "Failed to align and sort seuqnece reads, aborting"
+    exit 1
+fi
 
 # Mark duplicate reads.
-{command1}
+{command2}
 
 # Index BAM file.
 samtools index {args.output}/temp/{r.Name}.sorted.markdup.bam
 
 # Build BQSR model.
-{command2}
+{command3}
 
 # Apply BQSR model.
-{command3}
+{command4}
 
 # Remove temporary files.
 {remove} {args.output}/temp/{r.Name}.metrics
@@ -236,6 +235,11 @@ samtools index {args.output}/temp/{r.Name}.sorted.markdup.bam
 {remove} {args.output}/temp/{r.Name}.sorted.markdup.bam
 {remove} {args.output}/temp/{r.Name}.sorted.markdup.bam.bai
 """)
+
+    if args.job is None:
+        jid = ''
+    else:
+        jid = args.job + '-'
 
     with open(f'{args.output}/shell/qsubme.sh', 'w') as f:
         f.write(
@@ -247,7 +251,6 @@ samples=({" ".join(df.Name)})
 
 for sample in ${{samples[@]}}
 do
-  qsub {args.qsub1} -S /bin/bash -e $p/log -o $p/log -N S1-$sample $p/shell/S1-$sample.sh
-  qsub {args.qsub2} -S /bin/bash -e $p/log -o $p/log -N S2-$sample -hold_jid S1-$sample $p/shell/S2-$sample.sh
+  qsub {args.qsub} -S /bin/bash -e $p/log -o $p/log -N {jid}S1-$sample $p/shell/S1-$sample.sh
 done
 """)